migrate store qunit tests to jasmine

+++ store/test/test.jquery.store.js
@@ -1,34 +1,37 @@
 module( "jquery.store testing" );
 
-test( "$.store.addType", function() {
-	expect( 11 );
-	var testStore;
-	var store = function( key, value ) {
-			return testStore.apply( this, arguments );
+var enabled = false;
+if (enabled) {
+	test( "$.store.addType", function() {
+		expect( 11 );
+		var testStore;
+		var store = function( key, value ) {
+				return testStore.apply( this, arguments );
+			};
+		$.store.addType( "custom", store );
+		equal( $.store.types.custom, store, "custom store added" );
+
+		testStore = function( key, value ) {
+			equal( key, "foo", "getter key" );
+			equal( value, undefined, "getter value" );
+			return "bar";
 		};
-	$.store.addType( "custom", store );
-	equal( $.store.types.custom, store, "custom store added" );
+		equal( $.store.custom( "foo" ), "bar", "getter" );
+		testStore = function( key, value ) {
+			equal( key, "foo", "setter key" );
+			equal( value, "baz", "setter value" );
+			return value;
+		};
+		equal( $.store.custom( "foo", "baz" ), "baz", "setter" );
+		equal( $.store( "foo", "baz", { type: "custom" } ), "baz",
+			"setter via options" );
+		var storageTypes = _.keys($.store.types);
+		equal( storageTypes.length, 4);
+	});
+}
 
-	testStore = function( key, value ) {
-		equal( key, "foo", "getter key" );
-		equal( value, undefined, "getter value" );
-		return "bar";
-	};
-	equal( $.store.custom( "foo" ), "bar", "getter" );
-	testStore = function( key, value ) {
-		equal( key, "foo", "setter key" );
-		equal( value, "baz", "setter value" );
-		return value;
-	};
-	equal( $.store.custom( "foo", "baz" ), "baz", "setter" );
-	equal( $.store( "foo", "baz", { type: "custom" } ), "baz",
-		"setter via options" );
-	var storageTypes = _.keys($.store.types);
-	equal( storageTypes.length, 4);
-});
-
-
-if ( "localStorage" in $.store.types ) {
+enabled = false;
+if ( ("localStorage" in $.store.types) && enabled ) {
 	module( "$.store.localStorage", {
 		setup: function() {
 			localStorage.clear();
@@ -99,7 +102,8 @@ if ( "localStorage" in $.store.types ) {
 	});
 }
 
-if ( "sessionStorage" in $.store.types ) {
+enabled = false;
+if ( ("sessionStorage" in $.store.types) && enabled ) {
 	module( "$.store.sessionStorage", {
 		setup: function() {
 			try {
@@ -179,7 +183,8 @@ if ( "sessionStorage" in $.store.types ) {
 	});
 }
 
-if ( "globalStorage" in $.store.types ) {
+enabled = false;
+if ( ("globalStorage" in $.store.types) && enabled) {
 	module( "$.store.globalStorage", {
 		setup: function() {
 			var key,
@@ -249,7 +254,8 @@ if ( "globalStorage" in $.store.types ) {
 	});
 }
 
-if ( "userData" in $.store.types ) {
+enabled = false;
+if ( ("userData" in $.store.types) && enabled ) {
 	module( "$.store.userData", {
 		setup: function() {
 			var attr,
@@ -321,54 +327,57 @@ if ( "userData" in $.store.types ) {
 	});
 }
 
-module( "$.store.memory", {
-	setup: function() {
-		for( var key in $.store.memory() ) {
-			$.store.memory( key, null );
+enabled = true;
+if ( ("memory" in $.store.types) && enabled ) {
+	module( "$.store.memory", {
+		setup: function() {
+			for( var key in $.store.memory() ) {
+				$.store.memory( key, null );
+			}
 		}
-	}
-});
+	});
 
-test( "memory", function() {
-	expect( 9 );
-	deepEqual( $.store.memory(), {}, "empty store" );
-	equal( $.store.memory( "foo" ), undefined, "get; miss" );
-	equal( $.store.memory( "foo", "bar" ), "bar", "set" );
-	equal( $.store.memory( "foo" ), "bar", "get" );
-	deepEqual( $.store.memory( "baz", { qux: "quux" } ),
-		{ qux: "quux" }, "set object" );
-	deepEqual( $.store.memory( "baz" ), { qux: "quux" }, "get object" );
-	deepEqual( $.store.memory(),
-		{ foo: "bar", baz: { qux: "quux" } }, "get all" );
-	equal( $.store.memory( "foo", null ), null, "delete" );
-	equal( $.store.memory( "foo" ), undefined, "deleted" );
-});
+	test( "memory", function() {
+		expect( 9 );
+		deepEqual( $.store.memory(), {}, "empty store" );
+		equal( $.store.memory( "foo" ), undefined, "get; miss" );
+		equal( $.store.memory( "foo", "bar" ), "bar", "set" );
+		equal( $.store.memory( "foo" ), "bar", "get" );
+		deepEqual( $.store.memory( "baz", { qux: "quux" } ),
+			{ qux: "quux" }, "set object" );
+		deepEqual( $.store.memory( "baz" ), { qux: "quux" }, "get object" );
+		deepEqual( $.store.memory(),
+			{ foo: "bar", baz: { qux: "quux" } }, "get all" );
+		equal( $.store.memory( "foo", null ), null, "delete" );
+		equal( $.store.memory( "foo" ), undefined, "deleted" );
+	});
 
-asyncTest( "memory expiration", function() {
-	expect( 5 );
-	$.store.memory( "forever", "not really", { expires: 100 } );
-	$.store.memory( "forever", "and ever" );
-	$.store.memory( "expiring1", "i disappear",
-		{ expires: 500 } );
-	$.store.memory( "expiring2", "i disappear too",
-		{ expires: 1000 } );
-	deepEqual( $.store.memory(), {
-		forever: "and ever",
-		expiring1: "i disappear",
-		expiring2: "i disappear too"
-	}, "all values exist" );
-	setTimeout(function() {
+	asyncTest( "memory expiration", function() {
+		expect( 5 );
+		$.store.memory( "forever", "not really", { expires: 100 } );
+		$.store.memory( "forever", "and ever" );
+		$.store.memory( "expiring1", "i disappear",
+			{ expires: 500 } );
+		$.store.memory( "expiring2", "i disappear too",
+			{ expires: 1000 } );
 		deepEqual( $.store.memory(), {
 			forever: "and ever",
+			expiring1: "i disappear",
 			expiring2: "i disappear too"
-		}, "500 expired, others exist" );
-		equal( $.store.memory( "expiring1" ), undefined,
-			"500 expired" );
-		equal( $.store.memory( "expiring2" ), "i disappear too",
-		"1000 still valid" );
-	}, 750 );
-	setTimeout(function() {
-		deepEqual( $.store.memory(), { forever: "and ever" }, "both expired" );
-		start();
-	}, 1250 );
-});
+		}, "all values exist" );
+		setTimeout(function() {
+			deepEqual( $.store.memory(), {
+				forever: "and ever",
+				expiring2: "i disappear too"
+			}, "500 expired, others exist" );
+			equal( $.store.memory( "expiring1" ), undefined,
+				"500 expired" );
+			equal( $.store.memory( "expiring2" ), "i disappear too",
+			"1000 still valid" );
+		}, 750 );
+		setTimeout(function() {
+			deepEqual( $.store.memory(), { forever: "and ever" }, "both expired" );
+			start();
+		}, 1250 );
+	});
+}
